refactor(button-menus): props to types
