+++ app/screens/Menu/ChangePasswordSuccess/ChangePasswordSuccess.types.ts
