+++ app/screens/Menu/FrequentlyAskedQuestions/FrequentlyAskedQuestions.types.ts
