+++ app/screens/Menu/FrequentlyAskedQuestions/components/FAQSectionList/FAQSectionList.types.ts
